refactor: remove unused WSDL version translation exception and clean up imports
